screen-cast-stream-src: Split up maybe_record API
Split out the record part and introduce a API variant that doesn't try
to throttle. This will be useful when the throttling happens consumer
side, when the consumer is the driver, or when throttling happens
implicitly by configuring the frame clock.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4398>