screen-cast-stream-src/virtual: Stop inhibiting hw cursor
The hardware cursor support in the cursor renderer now acts on a per
stage view basis, so no need to inhibit the hardware cursor, it isn't
going to try to put a hardware cursor on a virtual CRTC anyway.

Instead use overlay inhibitation to decide whether an cursor overlay
should be painted or not. Inhibit the overlay when hidden/metadata, but
not when embedded.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4066>