screen-cast-stream-src/virtual: Never "maybe" record frames
The frame clock is configured from the negotiated framerate, so we
should never end up in a situation where we draw to fast, so there is no
point in doing the additional frame rate that e.g. monitor sources need
to do.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4398>