dbus-session-watcher: Add API to queue closing in an idle callback
This is useful if we're closing while dispatching a PipeWire source.
While this is unlikely to happen right now, when we drive the frame
clock from PipeWire graph processing, it might become a reality, and by
then we'll need to close outside of the current calling context.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4398>