ci: Add pkgconf wrapper in install-common-dependencies
The script is used to install dependencies that are needed both
for the CI image and the gnome os system extension.

To avoid installing dependencies in an environment where they are
already satisfied, they should be guarded by an appropriate check,
most commonly pkgconf.

Right now these checks are straight-forward:
 - if a dependency is not in the base image, it will
   be built when building the CI image
 - if a dependency is not in GNOME OS, it will be
   built when building the GNOME OS system extension

However we will soon add another caller of the script, when
allowing to export a system extension from our development
toolboxes to the host.

That case is trickier, as the dependency *is* satisfied by
the build environment (because it is based on the CI image),
but it will still be needed by the exported system extension.

To account for that case, add a pkgconf wrapper that only searches
in PKG_CONFIG_DIRS in the specified destdirs.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4075>