clutter/color-state: Introduce ClutterColorStateIcc
This new color-state parses an ICC profile using lcms2.

When parsing it, generates eotf/inv_eotf profiles that will be used if
the transform is required to go from linear / to linear. This allows
blending in linear.

The transform snippet is always the same: a 3D LUT with tetrahedral
interpolation.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4102>