backend: Reset panel rotation when orientation is no longer managed
When the panel orientation managed state goes from managed to unmanaged, we
stop listening to orientation sensor changes. This means we won't receive an
event when the orientation of the device goes back to normal, and the screen
will be stuck in a "non-normal" rotation.

Avoid that and reset the panel back to the normal rotation right away when
panel_orientation_managed gets set to FALSE.

To test this new behavior, adapt the intial-stored-rotated MonitorUnit test.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>