orientation-manager: Add API for inhibiting orientation change listening
As promised in the last commit, introduce inhibit/uninhibit API for the
orientation-manager. This is quite straightforward now after the refactor.

We use this API in the orientation manager itself for the rotation
lock, and we'll use it in the subsequent commits from the monitor manager to
inhibit chages when the panel orientation isn't managed.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>