orientation-manager: Refactor orientation changed idle a bit
Delay the actual "orientation changed" handler using an idle, rather than
the delaying the sync_state() function. This will be helpful in the next
commit, where we move the has_accel state outside of sync_state().

Note that this means we no longer delay the sync_state() after changes to
the rotation lock, and rather do that immediately. That should not be an
issue though, as the rotation lock shouldn't change while the session is
inactive (which is what commit 4fe75695 ("orientation-manager: Use an idle
to apply monitor orientation changes") was working around).

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>