backend: Inhibit orientation sensor when panel orientation is not managed
There's no need to power up the accelerometer when the panel orientation is
not managed, so use the new (un)inhibit APIs from MetaOrientationManager to
make that happen.

In order for the "initial orientation change workaround for portrait mode
panels" to continue working, we can't inhibit the sensor right away when
starting up, but we must wait with that until the first orientation
measurement was received.

Because the orientation manager now no longer picks up rotation changes that
happen without the sensor being claimed, also adapt the orientation tests
and remove the meta_wait_for_orientation() calls in those cases. Instead we
can now assert that mutter isn't claiming the sensor.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>