wayland/surface: Fix bounding box rect when calculating scanout
The "dst_rect" calculated is in "CRTC space", meaning the bounding box
used for calculating it should be the view layout dimension, scaled by
the view scale and transform by the view transform. Previously it was
only transformed, not scaled. While fixing this, rename the variables to
make it a bit more clear what coordinate space they are expected to be
in.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4147>