clutter: Move sprite creation closer to event handling entrypoint
Have the sprite be created in the only entrypoint that could create
it, and avoid doing that in the lower layers. This will allow moving
some more code to ClutterSprite, and make more code in ClutterStage
deal with ClutterSprite only, as opposed to devices and sequences.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4193>