screenshot: Avoid using cairo in public API
Currently, the window screenshoting API writes a CoglTexture's data into
the system memory and exposes it as a cairo_surface_t in the public API
as that what GNOME Shell uses afterwards.

Although, that is nice, returning a CoglTexture instead would make us
get rid of cairo in the public API and along with the patches that makes
the font renderning optional if they ever land and another patch to make
the cursor drawing for screencasting not use cairo, one could have a
cairo-less build when building without font rendering & x11.