backends/input-mapper: Synchronize device enabled state when adding device
When device is added to the mapper, the device may not be in a
desirable state. This changeset updates the device state when
the device is added to the input-mapper.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3834
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4188>
(cherry picked from commit 78c621aa1acc248b0c62a46a925160f53dda5ed7)