compositor: Add vmethod to handle an event
And integrate it in core/events.c event handling as a last
step in event processing. We will want to handle things that
are dependent of the backend, and the MetaCompositor implementation
will have a more direct knowledge of the backend implementation it
duos with.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4191>