core: Move releasing of passive button grabs to X11 compositor
This is something we only need to do running as a X11 WM/compositor,
and the MetaCompositorX11 implementation is forcibly paired with a
MetaBackendX11Cm.

Handle the XIAllowEvent calls through the handle_event vmethod there,
and drop this piece of X11-specific code from src/core.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4191>