backends/x11: Add methods to create/destroy passive key grabs
Add similar API to the existing one for button grabs, specific to the
MetaBackendX11 implementation. The same reasoning as there is in button
grabs applies here, WRT locked modifiers being handled out of the box
in XInput2. Thus we do not need to calculate the combination of modifiers
containing lock keys.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4191>