compositor/x11: Take over issuing passive key grabs
Handle passive key grabs for all keybindings on client windows
and the root window at MetaCompositorX11, ensuring preferences
are honored.

In order to do this, some per-window state tracking has to be
performed on window type and decorated changes, in order to
ensure only windows of the expected type get keygrabs, and these
keygrabs happen always on the toplevel window.

This drops a bit of code in src/core and src/x11 that were actually
specific to the c-m X11 backend.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4191>