tests/cursor: Set cursor theme to Adwaita
The reference images were generated using the Adwaita theme, so make sure that
the test client also draws a cursrom from the Adwaita theme rather than reading the
default theme from the environment.

Ideally the test suite would embed the cursor theme and point XCURSOR_PATH to
it, so that even if the Adwaita icon theme changes in the future the test
wouldn't suddently break; I did not implement this yet.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4225>