backend/native: Don't create launcher for headless, require for others
Headless sessions don't get associated with the session. They are not
supposed to get the devices of the session, nor track the session.
Simply never create a launcher for them.

For the test sessions we want to be able to test things involving
devices, so create a launcher for them so and let the mocked logind take
care of it.

The default mode does require a launcher.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4448>